+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1-alpha2] - 2021-06-19
+### Added
+- Add checkbox option to close main window when stopping a task when invoked from a tray
+- Add elapsed time counter to main window when a task is running
+
+### Changed
+- Trim the active task text, so it doesn't expand the window size
+
+### Removed
+- Removed QUIT button from main window
+
 ## [0.3.1-alpha1] - 2021-06-19
 ### Added
 - Add checkbox option to close main window when creating and starting a new task when invoked from the tray